test(instance-of): use instanceof instead of typeof
This commit updates a test suite to use instanceof instead of typeof to check the class type of an object.

This change is made because instanceof provides a more accurate check of the object's class type than typeof. Typeof will return "object" for all objects, regardless of their class, which can lead to incorrect results. Instanceof, on the other hand, will return true only if the object is an instance of the specified class.